InfrrdTask-1/requirements.txt
flask==2.2.2
flask-restful==0.3.9
pymongo==3.12.1
python-dotenv==0.19.2
werkzeug==2.2.2
pandas==1.5.3
bottleneck==1.3.6
sentence-transformers==2.2.0
tensorflow==2.13.0
